fix: image torch encoders must implement encode
