+++ jina/executors/encoders/image/paddlehub.py
@@ -1,7 +1,10 @@
 __copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
+import numpy as np
+
 from ..frameworks import BasePaddleEncoder
+from ...decorators import batching, as_ndarray
 
 
 class ImagePaddlehubEncoder(BasePaddleEncoder):
@@ -12,7 +15,13 @@ class ImagePaddlehubEncoder(BasePaddleEncoder):
     https://github.com/PaddlePaddle/PaddleHub
     """
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self,
+                 model_name: str = None,
+                 output_feature: str = None,
+                 pool_strategy: str = None,
+                 channel_axis: int = -3,
+                 *args,
+                 **kwargs):
         """
 
         :param model_name: the name of the model. Supported models include
@@ -41,6 +50,12 @@ class ImagePaddlehubEncoder(BasePaddleEncoder):
 
         """
         super().__init__(*args, **kwargs)
+        self.pool_strategy = pool_strategy
+        self.outputs_name = output_feature
+        self.inputs_name = None
+        self.channel_axis = channel_axis
+        self._default_channel_axis = -3
+        self.model_name = model_name
         if self.model_name is None:
             self.model_name = 'xception71_imagenet'
         if self.outputs_name is None:
@@ -48,8 +63,44 @@ class ImagePaddlehubEncoder(BasePaddleEncoder):
         if self.pool_strategy is None:
             self.pool_strategy = 'mean'
 
+    def post_init(self):
+        super().post_init()
+        import paddlehub as hub
+        module = hub.Module(name=self.model_name)
+        inputs, outputs, self.model = module.context(trainable=False)
+        self.get_inputs_and_outputs_name(inputs, outputs)
+        self.exe = self.to_device()
+
+    def close(self):
+        self.exe.close()
+
     def get_inputs_and_outputs_name(self, input_dict, output_dict):
         self.inputs_name = input_dict['image'].name
         self.outputs_name = output_dict['feature_map'].name
         if self.model_name.startswith('vgg') or self.model_name.startswith('alexnet'):
             self.outputs_name = f'@HUB_{self.model_name}@fc_1.tmp_2'
+
+    @batching
+    @as_ndarray
+    def encode(self, data: 'np.ndarray', *args, **kwargs) -> 'np.ndarray':
+        """
+
+        :param data: a `B x T x (Channel x Height x Width)` numpy ``ndarray``, `B` is the size of the batch, `T` is the
+            number of frames
+        :return: a `B x D` numpy ``ndarray``, `D` is the output dimension
+        """
+        if self.channel_axis != self._default_channel_axis:
+            data = np.moveaxis(data, self.channel_axis, self._default_channel_axis)
+        feature_map, *_ = self.exe.run(
+            program=self.model,
+            fetch_list=[self.outputs_name],
+            feed={self.inputs_name: data.astype('float32')},
+            return_numpy=True
+        )
+        if feature_map.ndim == 2 or self.pool_strategy is None:
+            return feature_map
+        return self.get_pooling(feature_map)
+
+    def get_pooling(self, data: 'np.ndarray') -> 'np.ndarray':
+        _reduce_axis = tuple((i for i in range(len(data.shape)) if i > 1))
+        return getattr(np, self.pool_strategy)(data, axis=_reduce_axis)
