+++ jina/executors/encoders/image/torchvision.py
@@ -4,6 +4,7 @@ __license__ = "Apache-2.0"
 import numpy as np
 
 from ..frameworks import BaseTorchEncoder
+from ...decorators import batching, as_ndarray
 
 
 class ImageTorchEncoder(BaseTorchEncoder):
@@ -14,7 +15,10 @@ class ImageTorchEncoder(BaseTorchEncoder):
     https://pytorch.org/docs/stable/torchvision/models.html
     """
 
-    def __init__(self, model_name: str = None, pool_strategy: str = 'mean', *args, **kwargs):
+    def __init__(self, model_name: str = None,
+                 pool_strategy: str = 'mean',
+                 channel_axis: int = 1,
+                 *args, **kwargs):
         """
 
         :param model_name: the name of the model. Supported models include
@@ -37,6 +41,8 @@ class ImageTorchEncoder(BaseTorchEncoder):
             - `max` means that global max pooling will be applied.
         """
         super().__init__(*args, **kwargs)
+        self.channel_axis = channel_axis
+        self._default_channel_axis = 1
         self.model_name = 'mobilenet_v2' or model_name
         if pool_strategy not in ('mean', 'max', None):
             raise NotImplementedError(f'unknown pool_strategy: {self.pool_strategy}')
@@ -58,6 +64,21 @@ class ImageTorchEncoder(BaseTorchEncoder):
         if feature_map.ndim == 2 or self.pool_strategy is None:
             return feature_map
         return self.pool_fn(feature_map, axis=(2, 3))
+
+    @batching
+    @as_ndarray
+    def encode(self, data: 'np.ndarray', *args, **kwargs) -> 'np.ndarray':
+        if self.channel_axis != self._default_channel_axis:
+            data = np.moveaxis(data, self.channel_axis, self._default_channel_axis)
+        import torch
+        _input = torch.from_numpy(data.astype('float32'))
+        if self.on_gpu:
+            _input = _input.cuda()
+        _feature = self._get_features(_input).detach()
+        if self.on_gpu:
+            _feature = _feature.cpu()
+        _feature = _feature.numpy()
+        return self._get_pooling(_feature)
 
 
 class CustomImageTorchEncoder(ImageTorchEncoder):
